Moved strategies to 1 place Namespace fixes
