+++ GhostfolioAPI/BalanceCalculator.cs
@@ -1,7 +1,6 @@
 ﻿using GhostfolioSidekick.Model;
 using GhostfolioSidekick.Model.Accounts;
 using GhostfolioSidekick.Model.Activities;
-using GhostfolioSidekick.Model.Activities.Types;
 using GhostfolioSidekick.Model.Compare;
 
 namespace GhostfolioSidekick.GhostfolioAPI
