+++ GhostfolioAPI/Contract/BalanceList.cs
@@ -1,10 +1,4 @@
-﻿using System;
-using System.Collections.Generic;
-using System.Linq;
-using System.Text;
-using System.Threading.Tasks;
-
-namespace GhostfolioSidekick.GhostfolioAPI.Contract
+﻿namespace GhostfolioSidekick.GhostfolioAPI.Contract
 {
 	public class BalanceList
 	{
