+++ GhostfolioAPI/Strategies/NotNativeSupportedTransactionsInGhostfolio.cs
@@ -1,7 +1,6 @@
 ﻿using GhostfolioSidekick.Model;
 using GhostfolioSidekick.Model.Activities;
 using GhostfolioSidekick.Model.Activities.Types;
-using GhostfolioSidekick.Model.Strategies;
 
 namespace GhostfolioSidekick.GhostfolioAPI.Strategies
 {
