+++ IntegrationTests/Run.cs
@@ -1,11 +1,10 @@
 using FluentAssertions;
-using GhostfolioSidekick;
 using GhostfolioSidekick.GhostfolioAPI;
 using Microsoft.Extensions.DependencyInjection;
 using Microsoft.Extensions.Hosting;
 using Microsoft.Extensions.Logging;
 
-namespace IntegrationTests
+namespace GhostfolioSidekick.IntegrationTests
 {
 	public class Run
 	{
